+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -294,6 +294,8 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
       keepPrimaryContainer: controller.keepDarkPrimaryContainer,
       keepSecondaryContainer: controller.keepDarkSecondaryContainer,
       keepTertiaryContainer: controller.keepDarkTertiaryContainer,
+      noOnMainsTint: controller.noDarkOnMainsTint,
+      noOnSurfacesTint: controller.noDarkOnSurfacesTint,
     ),
     useMaterial3ErrorColors: controller.useM3ErrorColors,
     tones: FlexTone.values[controller.usedFlexToneSetup].tones(Brightness.dark),