+++ example/lib/example5_themes_playground/widgets/panels/seeded_color_scheme_settings/scheme_colors.dart
@@ -61,8 +61,11 @@ class SchemeColors extends StatelessWidget {
     final Color tertiaryContainer = colorScheme.tertiaryContainer;
 
     // Grab active tons and chroma setup.
-    final FlexTones tones =
-        FlexTone.values[tc.usedFlexToneSetup].tones(theme.brightness);
+    final FlexTones tones = FlexTone.values[tc.usedFlexToneSetup]
+        .tones(theme.brightness)
+        .noOnMainsTint(isLight ? tc.noLightOnMainsTint : tc.noDarkOnMainsTint)
+        .noOnSurfacesTint(
+            isLight ? tc.noLightOnSurfacesTint : tc.noDarkOnSurfacesTint);
 
     // Should we even show the tone? We show them only when, seeding is on.
     final bool showTones = tc.useKeyColors && tc.useFlexColorScheme;