+++ test/flex_key_color_test.dart
@@ -19,6 +19,8 @@ void main() {
       keepPrimaryContainer: false,
       keepSecondaryContainer: false,
       keepTertiaryContainer: false,
+      noOnMainsTint: false,
+      noOnSurfacesTint: false,
     );
     // m2, has same definition as m1, but via default values
     const FlexKeyColors m2 = FlexKeyColors();
@@ -33,6 +35,8 @@ void main() {
       keepPrimaryContainer: false,
       keepSecondaryContainer: false,
       keepTertiaryContainer: false,
+      noOnMainsTint: false,
+      noOnSurfacesTint: false,
     );
     // m4, has all values different from m1
     const FlexKeyColors m4 = FlexKeyColors(
@@ -45,6 +49,8 @@ void main() {
       keepPrimaryContainer: true,
       keepSecondaryContainer: true,
       keepTertiaryContainer: true,
+      noOnMainsTint: true,
+      noOnSurfacesTint: true,
     );
     // Do identity tests
     test(
@@ -119,6 +125,8 @@ void main() {
           keepPrimaryContainer: false,
           keepSecondaryContainer: false,
           keepTertiaryContainer: false,
+          noOnMainsTint: false,
+          noOnSurfacesTint: false,
         ),
         equals(m1),
       );
@@ -138,7 +146,7 @@ void main() {
           //
           equalsIgnoringHashCodes(
               // ignore: lines_longer_than_80_chars
-              'FlexKeyColors#00000(useKeyColors: true, useSecondary: false, useTertiary: false, keepPrimary: false, keepSecondary: false, keepTertiary: false, keepPrimaryContainer: false, keepSecondaryContainer: false, keepTertiaryContainer: false)'));
+              'FlexKeyColors#00000(useKeyColors: true, useSecondary: false, useTertiary: false, keepPrimary: false, keepSecondary: false, keepTertiary: false, keepPrimaryContainer: false, keepSecondaryContainer: false, keepTertiaryContainer: false, noOnMainsTint: false, noOnSurfacesTint: false)'));
     });
     test(
         'FKCS1.11: Test toStringShort implemented via debugFillProperties '